Refactoring powdr asm generator to make it source agnostic.
