+++ riscv/benches/executor_benchmark.rs
@@ -2,7 +2,8 @@ use ::powdr_pipeline::Pipeline;
 use powdr_number::GoldilocksField;
 
 use powdr_riscv::{
-    compile_rust_crate_to_riscv_asm, compiler, continuations::bootloader::default_input, Runtime,
+    asm_translate, compile_rust_crate_to_riscv_asm, continuations::bootloader::default_input,
+    Runtime,
 };
 
 use criterion::{criterion_group, criterion_main, Criterion};
@@ -18,7 +19,7 @@ fn executor_benchmark(c: &mut Criterion) {
     let tmp_dir = Temp::new_dir().unwrap();
     let riscv_asm_files =
         compile_rust_crate_to_riscv_asm("./tests/riscv_data/keccak/Cargo.toml", &tmp_dir);
-    let contents = compiler::compile::<T>(riscv_asm_files, &Runtime::base(), false);
+    let contents = asm_translate::compile::<T>(riscv_asm_files, &Runtime::base(), false);
     let mut pipeline = Pipeline::<T>::default().from_asm_string(contents, None);
     pipeline.compute_optimized_pil().unwrap();
     pipeline.compute_fixed_cols().unwrap();
@@ -30,7 +31,8 @@ fn executor_benchmark(c: &mut Criterion) {
     // The first chunk of `many_chunks`, with Poseidon co-processor & bootloader
     let riscv_asm_files =
         compile_rust_crate_to_riscv_asm("./tests/riscv_data/many_chunks/Cargo.toml", &tmp_dir);
-    let contents = compiler::compile::<T>(riscv_asm_files, &Runtime::base().with_poseidon(), true);
+    let contents =
+        asm_translate::compile::<T>(riscv_asm_files, &Runtime::base().with_poseidon(), true);
     let mut pipeline = Pipeline::<T>::default().from_asm_string(contents, None);
     pipeline.compute_optimized_pil().unwrap();
     pipeline.compute_fixed_cols().unwrap();
