+++ riscv/src/code_gen.rs
@@ -1,27 +1,12 @@
-use std::{
-    collections::{BTreeMap, BTreeSet, HashSet},
-    fmt,
-};
+use std::fmt;
 
 use itertools::Itertools;
-use powdr_asm_utils::{
-    ast::{BinaryOpKind, UnaryOpKind},
-    data_parser,
-    data_storage::{store_data_objects, SingleDataValue},
-    parser::parse_asm,
-    reachability::{self, symbols_in_args},
-    utils::{
-        argument_to_escaped_symbol, argument_to_number, escape_label, expression_to_number, quote,
-    },
-    Architecture,
-};
+use powdr_asm_utils::data_storage::SingleDataValue;
+use powdr_asm_utils::utils::{escape_label, quote};
 use powdr_number::{FieldElement, KnownField};
 
 use crate::continuations::bootloader::{bootloader_and_shutdown_routine, bootloader_preamble};
-use crate::disambiguator;
-use crate::parser::RiscParser;
 use crate::runtime::Runtime;
-use crate::{Argument, Expression, Statement};
 
 #[derive(Clone, Copy, PartialEq, Eq, Debug)]
 pub struct Register {
@@ -63,135 +48,115 @@ impl fmt::Display for FunctionKind {
     }
 }
 
-struct RiscvArchitecture {}
-
-impl Architecture for RiscvArchitecture {
-    fn instruction_ends_control_flow(instr: &str) -> bool {
-        match instr {
-            "li" | "lui" | "la" | "mv" | "add" | "addi" | "sub" | "neg" | "mul" | "mulh"
-            | "mulhu" | "mulhsu" | "divu" | "remu" | "xor" | "xori" | "and" | "andi" | "or"
-            | "ori" | "not" | "slli" | "sll" | "srli" | "srl" | "srai" | "seqz" | "snez"
-            | "slt" | "slti" | "sltu" | "sltiu" | "sgtz" | "beq" | "beqz" | "bgeu" | "bltu"
-            | "blt" | "bge" | "bltz" | "blez" | "bgtz" | "bgez" | "bne" | "bnez" | "jal"
-            | "jalr" | "call" | "ecall" | "ebreak" | "lw" | "lb" | "lbu" | "lh" | "lhu" | "sw"
-            | "sh" | "sb" | "nop" | "fence" | "fence.i" | "amoadd.w" | "amoadd.w.aq"
-            | "amoadd.w.rl" | "amoadd.w.aqrl" | "lr.w" | "lr.w.aq" | "lr.w.rl" | "lr.w.aqrl"
-            | "sc.w" | "sc.w.aq" | "sc.w.rl" | "sc.w.aqrl" => false,
-            "j" | "jr" | "tail" | "ret" | "unimp" => true,
-            _ => {
-                panic!("Unknown instruction: {instr}");
-            }
-        }
-    }
-
-    fn get_references<
-        'a,
-        R: powdr_asm_utils::ast::Register,
-        F: powdr_asm_utils::ast::FunctionOpKind,
-    >(
-        instr: &str,
-        args: &'a [powdr_asm_utils::ast::Argument<R, F>],
-    ) -> Vec<&'a str> {
-        // fence arguments are not symbols, they are like reserved
-        // keywords affecting the instruction behavior
-        if instr.starts_with("fence") {
-            Vec::new()
-        } else {
-            symbols_in_args(args)
-        }
-    }
+pub enum Statement<'a, A: Args + ?Sized> {
+    DebugLoc { file: u64, line: u64, col: u64 },
+    Label(&'a str),
+    Instruction { op: &'a str, args: &'a A },
 }
 
-/// Compiles riscv assembly to a powdr assembly file. Adds required library routines.
-pub fn compile<T: FieldElement>(
-    mut assemblies: BTreeMap<String, String>,
+pub struct MemEntry {
+    pub label: Option<String>,
+    pub addr: u32,
+    pub value: SingleDataValue,
+}
+
+pub struct SourceFileInfo<'a> {
+    pub id: u32,
+    pub dir: &'a str,
+    pub file: &'a str,
+}
+
+/// A RISC-V program that can be translated to POWDR ASM.
+pub trait RiscVProgram {
+    type InstructionArgs: Args + ?Sized;
+    type Label;
+
+    // Source files to be used by the debug statements.
+    fn source_files_info(&self) -> impl Iterator<Item = SourceFileInfo>;
+
+    fn initial_mem(&self) -> impl Iterator<Item = MemEntry>;
+    fn executable_statements(&self) -> impl Iterator<Item = Statement<Self::InstructionArgs>>;
+    fn start_function(&self) -> &str;
+}
+
+pub fn translate_program<F: FieldElement>(
+    program: &impl RiscVProgram,
     runtime: &Runtime,
     with_bootloader: bool,
 ) -> String {
-    // stack grows towards zero
-    let stack_start = 0x10000;
-    // data grows away from zero
-    let data_start = 0x10100;
+    // Do this in a separate function to avoid most of the code being generic on F.
+    let (initial_mem, instructions, degree) =
+        translate_program_impl(program, runtime, with_bootloader);
 
-    assert!(assemblies
-        .insert("__runtime".to_string(), runtime.global_declarations())
-        .is_none());
+    riscv_machine(
+        runtime,
+        degree,
+        &preamble::<F>(runtime, with_bootloader),
+        initial_mem,
+        instructions,
+    )
+}
 
-    // TODO remove unreferenced files.
-    let (mut statements, file_ids) = disambiguator::disambiguate(
-        assemblies
-            .into_iter()
-            .map(|(name, contents)| (name, parse_asm(RiscParser::default(), &contents)))
-            .collect(),
-    );
-    let mut data_sections = data_parser::extract_data_objects(&statements);
-
-    // Reduce to the code that is actually reachable from main
-    // (and the objects that are referred from there)
-    let data_labels = reachability::filter_reachable_from::<_, _, RiscvArchitecture>(
-        "__runtime_start",
-        &mut statements,
-        &mut data_sections,
-    );
-
-    // Replace dynamic references to code labels
-    replace_dynamic_label_references(&mut statements, &data_labels);
-
+fn translate_program_impl(
+    program: &impl RiscVProgram,
+    runtime: &Runtime,
+    with_bootloader: bool,
+) -> (Vec<String>, Vec<String>, u64) {
     let mut initial_mem = Vec::new();
     let mut data_code = Vec::new();
-    let data_positions =
-        store_data_objects(data_sections, data_start, &mut |label, addr, value| {
-            if let Some(label) = label {
-                let comment = format!(" // data {label}");
-                if with_bootloader && !matches!(value, SingleDataValue::LabelReference(_)) {
-                    &mut initial_mem
+    for MemEntry { label, addr, value } in program.initial_mem() {
+        if let Some(label) = label {
+            // This is a comment, so we don't need to escape the label.
+            let comment = format!(" // data {label}");
+            if with_bootloader && !matches!(value, SingleDataValue::LabelReference(_)) {
+                &mut initial_mem
+            } else {
+                &mut data_code
+            }
+            .push(comment);
+        }
+        match value {
+            SingleDataValue::Value(v) => {
+                if with_bootloader {
+                    // Instead of generating the data loading code, we store it
+                    // in the variable that will be used as the initial memory
+                    // snapshot, committed by the bootloader.
+                    initial_mem.push(format!("(0x{addr:x}, 0x{v:x})"));
                 } else {
-                    &mut data_code
-                }
-                .push(comment);
-            }
-            match value {
-                SingleDataValue::Value(v) => {
-                    if with_bootloader {
-                        // Instead of generating the data loading code, we store it
-                        // in the variable that will be used as the initial memory
-                        // snapshot, committed by the bootloader.
-                        initial_mem.push(format!("(0x{addr:x}, 0x{v:x})"));
-                    } else {
-                        // There is no bootloader to commit to memory, so we have to
-                        // load it explicitly.
-                        data_code.push(format!("mstore 0x{addr:x}, 0x{v:x};"));
-                    }
-                }
-                SingleDataValue::LabelReference(sym) => {
-                    // The label value is not known at this point, so we have to
-                    // load it via code, irrespectively of bootloader availability.
-                    //
-                    // TODO should be possible without temporary
-                    data_code.extend([
-                        format!("tmp1 <== load_label({});", escape_label(sym)),
-                        format!("mstore 0x{addr:x}, tmp1;"),
-                    ]);
-                }
-                SingleDataValue::Offset(_, _) => {
-                    unimplemented!();
-                    /*
-                    object_code.push(format!("addr <=X= 0x{pos:x};"));
-
-                    I think this solution should be fine but hard to say without
-                    an actual code snippet that uses it.
-
-                    // TODO should be possible without temporary
-                    object_code.extend([
-                        format!("tmp1 <== load_label({});", escape_label(a)),
-                        format!("tmp2 <== load_label({});", escape_label(b)),
-                        // TODO check if registers match
-                        "mstore wrap(tmp1 - tmp2);".to_string(),
-                    ]);
-                    */
+                    // There is no bootloader to commit to memory, so we have to
+                    // load it explicitly.
+                    data_code.push(format!("mstore 0x{addr:x}, 0x{v:x};"));
                 }
             }
-        });
+            SingleDataValue::LabelReference(sym) => {
+                // The label value is not known at this point, so we have to
+                // load it via code, irrespectively of bootloader availability.
+                //
+                // TODO should be possible without temporary
+                data_code.extend([
+                    format!("tmp1 <== load_label({});", escape_label(&sym)),
+                    format!("mstore 0x{addr:x}, tmp1;"),
+                ]);
+            }
+            SingleDataValue::Offset(_, _) => {
+                unimplemented!();
+                /*
+                object_code.push(format!("addr <=X= 0x{pos:x};"));
+
+                I think this solution should be fine but hard to say without
+                an actual code snippet that uses it.
+
+                // TODO should be possible without temporary
+                object_code.extend([
+                    format!("tmp1 <== load_label({});", escape_label(a)),
+                    format!("tmp2 <== load_label({});", escape_label(b)),
+                    // TODO check if registers match
+                    "mstore wrap(tmp1 - tmp2);".to_string(),
+                ]);
+                */
+            }
+        }
+    }
 
     let submachines_init = runtime.submachines_init();
     let bootloader_and_shutdown_routine_lines = if with_bootloader {
@@ -205,26 +170,42 @@ pub fn compile<T: FieldElement>(
         submachines_init
     };
 
-    let mut program: Vec<String> = file_ids
-        .into_iter()
-        .map(|(id, dir, file)| format!(".debug file {id} {} {};", quote(&dir), quote(&file)))
+    let mut statements: Vec<String> = program
+        .source_files_info()
+        .map(
+            |SourceFileInfo {
+                 id,
+                 dir,
+                 file: name,
+             }| { format!(".debug file {id} {} {};", quote(dir), quote(name)) },
+        )
         .chain(bootloader_and_shutdown_routine_lines)
         .collect();
     if !data_code.is_empty() {
-        program.push("x1 <== jump(__data_init);".to_string());
+        statements.push("x1 <== jump(__data_init);".to_string());
     }
-    program.extend([
-        format!("// Set stack pointer\nx2 <=X= {stack_start};"),
-        "x1 <== jump(__runtime_start);".to_string(),
+    statements.extend([
+        format!("x1 <== jump({});", program.start_function()),
         "return;".to_string(), // This is not "riscv ret", but "return from powdr asm function".
     ]);
-    program.extend(
-        substitute_symbols_with_values(statements, &data_positions)
-            .into_iter()
-            .flat_map(process_statement),
-    );
+    for s in program.executable_statements() {
+        match s {
+            Statement::DebugLoc { file, line, col } => {
+                statements.push(format!(".debug loc {file} {line} {col};"))
+            }
+            Statement::Label(l) => statements.push(format!("{}:", escape_label(l))),
+            Statement::Instruction { op, args } => {
+                let processed_instr = match process_instruction(op, args) {
+                    Ok(s) => s,
+                    Err(e) => panic!("Failed to process instruction '{op}'. {e}"),
+                };
+                statements.extend(processed_instr.into_iter().map(|s| "  ".to_string() + &s))
+            }
+        }
+    }
+
     if !data_code.is_empty() {
-        program.extend(
+        statements.extend(
         ["// This is the data initialization routine.\n__data_init:".to_string()].into_iter()
         .chain(data_code)
         .chain([
@@ -232,10 +213,10 @@ pub fn compile<T: FieldElement>(
                 .to_string(),
         ]));
     }
-    program.extend(runtime.ecall_handler());
+    statements.extend(runtime.ecall_handler());
 
     // The program ROM needs to fit the degree, so we use the next power of 2.
-    let degree = program.len().ilog2() + 1;
+    let degree = statements.len().ilog2() + 1;
     let degree = std::cmp::max(degree, 18);
     log::info!("Inferred degree 2^{degree}");
 
@@ -250,157 +231,7 @@ pub fn compile<T: FieldElement>(
     assert!((18..=20).contains(&degree));
     let degree = 1 << degree;
 
-    riscv_machine(
-        runtime,
-        degree,
-        &preamble::<T>(runtime, with_bootloader),
-        initial_mem,
-        program,
-    )
-}
-
-/// Replace certain patterns of references to code labels by
-/// special instructions. We ignore any references to data objects
-/// because they will be handled differently.
-fn replace_dynamic_label_references(statements: &mut Vec<Statement>, data_labels: &HashSet<&str>) {
-    /*
-    Find patterns of the form
-    lui	a0, %hi(LABEL)
-    addi	s10, a0, %lo(LABEL)
-    -
-    turn this into the pseudoinstruction
-    li s10, LABEL
-    which is then turned into
-
-    s10 <== load_label(LABEL)
-
-    It gets complicated by the fact that sometimes, labels
-    and debugging directives occur between the two statements
-    matching that pattern...
-    */
-    let instruction_indices = statements
-        .iter()
-        .enumerate()
-        .filter_map(|(i, s)| match s {
-            Statement::Instruction(_, _) => Some(i),
-            _ => None,
-        })
-        .collect::<Vec<_>>();
-
-    let mut to_delete = BTreeSet::default();
-    for (i1, i2) in instruction_indices.into_iter().tuple_windows() {
-        if let Some(r) =
-            replace_dynamic_label_reference(&statements[i1], &statements[i2], data_labels)
-        {
-            to_delete.insert(i1);
-            statements[i2] = r;
-        }
-    }
-
-    let mut i = 0;
-    statements.retain(|_| (!to_delete.contains(&i), i += 1).0);
-}
-
-fn replace_dynamic_label_reference(
-    s1: &Statement,
-    s2: &Statement,
-    data_labels: &HashSet<&str>,
-) -> Option<Statement> {
-    let Statement::Instruction(instr1, args1) = s1 else {
-        return None;
-    };
-    let Statement::Instruction(instr2, args2) = s2 else {
-        return None;
-    };
-    if instr1.as_str() != "lui" || instr2.as_str() != "addi" {
-        return None;
-    };
-    let [Argument::Register(r1), Argument::Expression(Expression::FunctionOp(FunctionKind::HiDataRef, expr1))] =
-        &args1[..]
-    else {
-        return None;
-    };
-    // Maybe should try to reduce expr1 and expr2 before comparing deciding it is a pure symbol?
-    let Expression::Symbol(label1) = expr1.as_ref() else {
-        return None;
-    };
-    let [Argument::Register(r2), Argument::Register(r3), Argument::Expression(Expression::FunctionOp(FunctionKind::LoDataRef, expr2))] =
-        &args2[..]
-    else {
-        return None;
-    };
-    let Expression::Symbol(label2) = expr2.as_ref() else {
-        return None;
-    };
-    if r1 != r3 || label1 != label2 || data_labels.contains(label1.as_str()) {
-        return None;
-    }
-    Some(Statement::Instruction(
-        "li".to_string(),
-        vec![
-            Argument::Register(*r2),
-            Argument::Expression(Expression::Symbol(label1.clone())),
-        ],
-    ))
-}
-
-fn substitute_symbols_with_values(
-    mut statements: Vec<Statement>,
-    data_positions: &BTreeMap<String, u32>,
-) -> Vec<Statement> {
-    for s in &mut statements {
-        let Statement::Instruction(_name, args) = s else {
-            continue;
-        };
-        for arg in args {
-            arg.post_visit_expressions_mut(&mut |expression| match expression {
-                Expression::Number(_) => {}
-                Expression::Symbol(symb) => {
-                    if let Some(pos) = data_positions.get(symb) {
-                        *expression = Expression::Number(*pos as i64)
-                    }
-                }
-                Expression::UnaryOp(op, subexpr) => {
-                    if let Expression::Number(num) = subexpr.as_ref() {
-                        let result = match op {
-                            UnaryOpKind::BitwiseNot => !num,
-                            UnaryOpKind::Negation => -num,
-                        };
-                        *expression = Expression::Number(result);
-                    };
-                }
-                Expression::BinaryOp(op, subexprs) => {
-                    if let (Expression::Number(a), Expression::Number(b)) =
-                        (&subexprs[0], &subexprs[1])
-                    {
-                        let result = match op {
-                            BinaryOpKind::Or => a | b,
-                            BinaryOpKind::Xor => a ^ b,
-                            BinaryOpKind::And => a & b,
-                            BinaryOpKind::LeftShift => a << b,
-                            BinaryOpKind::RightShift => a >> b,
-                            BinaryOpKind::Add => a + b,
-                            BinaryOpKind::Sub => a - b,
-                            BinaryOpKind::Mul => a * b,
-                            BinaryOpKind::Div => a / b,
-                            BinaryOpKind::Mod => a % b,
-                        };
-                        *expression = Expression::Number(result);
-                    }
-                }
-                Expression::FunctionOp(op, subexpr) => {
-                    if let Expression::Number(num) = subexpr.as_ref() {
-                        let result = match op {
-                            FunctionKind::HiDataRef => num >> 12,
-                            FunctionKind::LoDataRef => num & 0xfff,
-                        };
-                        *expression = Expression::Number(result);
-                    };
-                }
-            });
-        }
-    }
-    statements
+    (initial_mem, statements, degree)
 }
 
 fn riscv_machine(
@@ -812,47 +643,8 @@ fn memory(with_bootloader: bool) -> String {
     "#
 }
 
-fn process_statement(s: Statement) -> Vec<String> {
-    match &s {
-        Statement::Label(l) => vec![format!("{}:", escape_label(l))],
-        Statement::Directive(directive, args) => match (directive.as_str(), &args[..]) {
-            (
-                ".loc",
-                [Argument::Expression(Expression::Number(file)), Argument::Expression(Expression::Number(line)), Argument::Expression(Expression::Number(column)), ..],
-            ) => {
-                vec![format!("  .debug loc {file} {line} {column};")]
-            }
-            (".file", _) => {
-                // We ignore ".file" directives because they have been extracted to the top.
-                vec![]
-            }
-            (".size", _) => {
-                // We ignore ".size" directives
-                vec![]
-            }
-            _ if directive.starts_with(".cfi_") => vec![],
-            _ => panic!(
-                "Leftover directive in code: {directive} {}",
-                args.iter().format(", ")
-            ),
-        },
-        Statement::Instruction(instr, args) => {
-            let stmt_str = format!("{s}");
-            // remove indentation and trailing newline
-            let stmt_str = &stmt_str[2..(stmt_str.len() - 1)];
-            let mut ret = vec![format!("  .debug insn \"{stmt_str}\";")];
-            let processed_instr = match process_instruction(instr, &args[..]) {
-                Ok(s) => s,
-                Err(e) => panic!("Failed to process instruction '{instr}'. {e}"),
-            };
-            ret.extend(processed_instr.into_iter().map(|s| "  ".to_string() + &s));
-            ret
-        }
-    }
-}
-
-trait Args {
-    type Error;
+pub trait Args {
+    type Error: fmt::Display;
 
     fn l(&self) -> Result<String, Self::Error>;
     fn r(&self) -> Result<Register, Self::Error>;
@@ -865,118 +657,6 @@ trait Args {
     fn rro(&self) -> Result<(Register, Register, u32), Self::Error>;
     fn rrro(&self) -> Result<(Register, Register, Register, u32), Self::Error>;
     fn empty(&self) -> Result<(), Self::Error>;
-}
-
-impl Args for [Argument] {
-    type Error = &'static str;
-
-    fn l(&self) -> Result<String, &'static str> {
-        const ERR: &str = "Expected: label";
-        match self {
-            [l] => Ok(argument_to_escaped_symbol(l).ok_or(ERR)?),
-            _ => Err(ERR),
-        }
-    }
-
-    fn r(&self) -> Result<Register, &'static str> {
-        match self {
-            [Argument::Register(r1)] => Ok(*r1),
-            _ => Err("Expected: register"),
-        }
-    }
-
-    fn rri(&self) -> Result<(Register, Register, u32), &'static str> {
-        const ERR: &str = "Expected: register, register, immediate";
-        match self {
-            [Argument::Register(r1), Argument::Register(r2), n] => {
-                Ok((*r1, *r2, argument_to_number(n).ok_or(ERR)?))
-            }
-            _ => Err(ERR),
-        }
-    }
-
-    fn rrr(&self) -> Result<(Register, Register, Register), &'static str> {
-        match self {
-            [Argument::Register(r1), Argument::Register(r2), Argument::Register(r3)] => {
-                Ok((*r1, *r2, *r3))
-            }
-            _ => Err("Expected: register, register, register"),
-        }
-    }
-
-    fn ri(&self) -> Result<(Register, u32), &'static str> {
-        const ERR: &str = "Expected: register, immediate";
-        match self {
-            [Argument::Register(r1), n] => Ok((*r1, argument_to_number(n).ok_or(ERR)?)),
-            _ => Err(ERR),
-        }
-    }
-
-    fn rr(&self) -> Result<(Register, Register), &'static str> {
-        match self {
-            [Argument::Register(r1), Argument::Register(r2)] => Ok((*r1, *r2)),
-            _ => Err("Expected: register, register"),
-        }
-    }
-
-    fn rrl(&self) -> Result<(Register, Register, String), &'static str> {
-        const ERR: &str = "Expected: register, register, label";
-        match self {
-            [Argument::Register(r1), Argument::Register(r2), l] => {
-                Ok((*r1, *r2, argument_to_escaped_symbol(l).ok_or(ERR)?))
-            }
-            _ => Err(ERR),
-        }
-    }
-
-    fn rl(&self) -> Result<(Register, String), &'static str> {
-        const ERR: &str = "Expected: register, label";
-        match self {
-            [Argument::Register(r1), l] => Ok((*r1, argument_to_escaped_symbol(l).ok_or(ERR)?)),
-            _ => Err(ERR),
-        }
-    }
-
-    fn rro(&self) -> Result<(Register, Register, u32), &'static str> {
-        if let [Argument::Register(r1), Argument::RegOffset(off, r2)] = self {
-            if let Some(off) = expression_to_number(off.as_ref().unwrap_or(&Expression::Number(0)))
-            {
-                return Ok((*r1, *r2, off));
-            }
-        }
-        if let [Argument::Register(r1), Argument::Expression(off)] = self {
-            if let Some(off) = expression_to_number(off) {
-                // If the register is not specified, it defaults to x0
-                return Ok((*r1, Register::new(0), off));
-            }
-        }
-
-        Err("Expected: register, offset(register)")
-    }
-
-    fn rrro(&self) -> Result<(Register, Register, Register, u32), &'static str> {
-        if let [Argument::Register(r1), Argument::Register(r2), Argument::RegOffset(off, r3)] = self
-        {
-            if let Some(off) = expression_to_number(off.as_ref().unwrap_or(&Expression::Number(0)))
-            {
-                return Ok((*r1, *r2, *r3, off));
-            }
-        }
-        if let [Argument::Register(r1), Argument::Register(r2), Argument::Expression(off)] = self {
-            if let Some(off) = expression_to_number(off) {
-                // If the register is not specified, it defaults to x0
-                return Ok((*r1, *r2, Register::new(0), off));
-            }
-        }
-        Err("Expected: register, register, offset(register)")
-    }
-
-    fn empty(&self) -> Result<(), &'static str> {
-        match self {
-            [] => Ok(()),
-            _ => Err("Expected: no arguments"),
-        }
-    }
 }
 
 fn only_if_no_write_to_zero(statement: String, reg: Register) -> Vec<String> {
