+++ riscv/tests/riscv.rs
@@ -22,7 +22,8 @@ pub fn test_continuations(case: &str) {
         &format!("tests/riscv_data/{case}/Cargo.toml"),
         &temp_dir,
     );
-    let powdr_asm = powdr_riscv::compiler::compile::<GoldilocksField>(riscv_asm, &runtime, true);
+    let powdr_asm =
+        powdr_riscv::asm_translate::compile::<GoldilocksField>(riscv_asm, &runtime, true);
 
     // Manually create tmp dir, so that it is the same in all chunks.
     let tmp_dir = mktemp::Temp::new_dir().unwrap();
@@ -274,7 +275,8 @@ fn many_chunks_dry() {
         &format!("tests/riscv_data/{case}/Cargo.toml"),
         &temp_dir,
     );
-    let powdr_asm = powdr_riscv::compiler::compile::<GoldilocksField>(riscv_asm, &runtime, true);
+    let powdr_asm =
+        powdr_riscv::asm_translate::compile::<GoldilocksField>(riscv_asm, &runtime, true);
 
     let mut pipeline = Pipeline::default()
         .from_asm_string(powdr_asm, Some(PathBuf::from(case)))
@@ -299,7 +301,8 @@ fn output_syscall() {
         &format!("tests/riscv_data/{case}/Cargo.toml"),
         &temp_dir,
     );
-    let powdr_asm = powdr_riscv::compiler::compile::<GoldilocksField>(riscv_asm, &runtime, false);
+    let powdr_asm =
+        powdr_riscv::asm_translate::compile::<GoldilocksField>(riscv_asm, &runtime, false);
 
     let inputs = vec![1u32, 2, 3]
         .into_iter()
@@ -373,5 +376,5 @@ fn compile_riscv_crate<T: FieldElement>(case: &str, runtime: &Runtime) -> String
         &format!("tests/riscv_data/{case}/Cargo.toml"),
         &temp_dir,
     );
-    powdr_riscv::compiler::compile::<T>(riscv_asm, runtime, false)
+    powdr_riscv::asm_translate::compile::<T>(riscv_asm, runtime, false)
 }
